# docs/contracts/openapi-v1-2026-02-19-phase5-dev.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "movement-whatsapp-automation-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/therapists": {
      "post": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Create Therapist",
        "description": "Create a new therapist (with associated User record).\n\nCreates both User and Therapist records atomically.",
        "operationId": "create_therapist_api_v1_admin_therapists_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TherapistCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "List Therapists",
        "description": "List all therapists with basic info.",
        "operationId": "list_therapists_api_v1_admin_therapists_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistListResponse"
                  },
                  "title": "Response List Therapists Api V1 Admin Therapists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/therapists/{therapist_id}": {
      "get": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Get Therapist",
        "description": "Get therapist detail with specialties.",
        "operationId": "get_therapist_api_v1_admin_therapists__therapist_id__get",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Update Therapist",
        "description": "Update therapist details.",
        "operationId": "update_therapist_api_v1_admin_therapists__therapist_id__patch",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TherapistUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Delete Therapist",
        "description": "Soft-delete therapist (set is_active=false).",
        "operationId": "delete_therapist_api_v1_admin_therapists__therapist_id__delete",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/therapist/patients": {
      "get": {
        "tags": [
          "Therapist Patients"
        ],
        "summary": "List Therapist Patients",
        "description": "List patients associated with the current therapist.",
        "operationId": "list_therapist_patients_api_v1_therapist_patients_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistPatientListItem"
                  },
                  "title": "Response List Therapist Patients Api V1 Therapist Patients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/therapist/patients/{client_id}": {
      "get": {
        "tags": [
          "Therapist Patients"
        ],
        "summary": "Get Therapist Patient Detail",
        "description": "Get patient detail scoped to the current therapist.",
        "operationId": "get_therapist_patient_detail_api_v1_therapist_patients__client_id__get",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistPatientDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/therapist/patients/{client_id}/sessions": {
      "get": {
        "tags": [
          "Therapist Patients"
        ],
        "summary": "List Therapist Patient Sessions",
        "description": "List sessions for a therapist-scoped patient.",
        "operationId": "list_therapist_patient_sessions_api_v1_therapist_patients__client_id__sessions_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistPatientSessionItem"
                  },
                  "title": "Response List Therapist Patient Sessions Api V1 Therapist Patients  Client Id  Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TherapistCreate": {
        "properties": {
          "neon_auth_sub": {
            "type": "string",
            "minLength": 1,
            "title": "Neon Auth Sub",
            "description": "Neon Auth subject ID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Therapist email (unique)"
          },
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri",
            "description": "Calendly user URI"
          }
        },
        "type": "object",
        "required": [
          "neon_auth_sub",
          "email",
          "display_name"
        ],
        "title": "TherapistCreate",
        "description": "Request schema for creating therapist.",
        "example": {
          "calendly_user_uri": "https://api.calendly.com/users/XXXXX",
          "display_name": "Dr. Smith",
          "email": "dr.smith@clinic.com",
          "license_number": "PT203315",
          "neon_auth_sub": "auth-therapist-123"
        }
      },
...
      "TherapistListResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "specialty_count": {
            "type": "integer",
            "title": "Specialty Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "license_number",
          "is_active",
          "email",
          "specialty_count"
        ],
        "title": "TherapistListResponse",
        "description": "Response schema for list of therapists (lightweight)."
      },
      "TherapistPatientDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone_e164": {
            "type": "string",
            "title": "Phone E164"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count"
          },
          "completed_session_count": {
            "type": "integer",
            "title": "Completed Session Count"
          },
          "upcoming_session_count": {
            "type": "integer",
            "title": "Upcoming Session Count"
          },
          "last_session_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session At"
          },
          "next_session_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Session At"
          },
          "plan_30": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssignedPlanSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_45": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssignedPlanSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone_e164",
          "email",
          "date_of_birth",
          "address",
          "session_count",
          "completed_session_count",
          "upcoming_session_count",
          "last_session_at",
          "next_session_at"
        ],
        "title": "TherapistPatientDetailResponse",
        "description": "Therapist-scoped patient detail."
      },
      "TherapistPatientListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone_e164": {
            "type": "string",
            "title": "Phone E164"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count"
          },
          "upcoming_session_count": {
            "type": "integer",
            "title": "Upcoming Session Count"
          },
          "last_session_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Session At"
          },
          "next_session_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Session At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone_e164",
          "email",
          "date_of_birth",
          "address",
          "session_count",
          "upcoming_session_count",
          "last_session_at",
          "next_session_at"
        ],
        "title": "TherapistPatientListItem",
        "description": "Therapist-scoped patient list item."
      },
...
      "TherapistResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri"
          },
          "specialties": {
            "items": {
              "$ref": "#/components/schemas/SpecialtyResponse"
            },
            "type": "array",
            "title": "Specialties"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "display_name",
          "license_number",
          "is_active",
          "calendly_user_uri",
          "specialties",
          "created_at"
        ],
        "title": "TherapistResponse",
        "description": "Response schema for therapist with full details."
      },
      "TherapistUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri"
          }
        },
        "type": "object",
        "title": "TherapistUpdate",
        "description": "Request schema for updating therapist."
      },
...
    }
  }
}